implement an example of lazy-loading (#3)
- lazy load the user state controller and template
- add a reference in README to lazy loading in the wiki documentation

As implemented, the lazy-loading employed here leverages webpack 2's support
for the pending dynamic import syntax. Eslint and babel now include plugins and
configurations that accept the new syntax.

Here's how it works: The dynamic import requests the files needed for the
application's routes. It provides the promise returned by the templateProvider
function for each route. For the controller, the approach is more indirect. At
the time of configuration, the route knows its controller by name, but that is
all. Within a resolve function, the ocLazyLoad service injects the module that
registers the controller. Now, when the router looks up the controller by name
it finds it in the registry.

Note: Grouping the related imports in a common chunk requires the use of a
"magic comment", which in turn required a webpack upgrade.